migrate resolve method to parent
